WalletScreen's component complete super finish WalletScreen

+++ Assets/1_Scripts/Models/ExpenseModel.cs
@@ -6,7 +6,7 @@ public class ExpenseModel
 {
     public int id;
     public string name;            
-    public decimal amount;           
+    public float amount;           
     public ExtraType type;           
     public DateTime? createdAt;     
 }
